Refactor resource notification handling
