+++ app/src/main/java/org/ole/planet/myplanet/ui/dashboard/DashboardViewModel.kt
@@ -3,17 +3,11 @@ package org.ole.planet.myplanet.ui.dashboard
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.realm.Realm
-import java.util.Date
-import java.util.UUID
 import javax.inject.Inject
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import org.ole.planet.myplanet.base.BaseResourceFragment
-import org.ole.planet.myplanet.datamanager.DatabaseService
-import org.ole.planet.myplanet.model.RealmNotification
 import org.ole.planet.myplanet.model.RealmSubmission
 import org.ole.planet.myplanet.repository.CourseRepository
 import org.ole.planet.myplanet.repository.LibraryRepository
@@ -23,12 +17,11 @@ import org.ole.planet.myplanet.repository.UserRepository
 
 @HiltViewModel
 class DashboardViewModel @Inject constructor(
-    private val databaseService: DatabaseService,
     private val userRepository: UserRepository,
     private val libraryRepository: LibraryRepository,
     private val courseRepository: CourseRepository,
     private val submissionRepository: SubmissionRepository,
-    private val notificationRepository: NotificationRepository
+    private val notificationRepository: NotificationRepository,
 ) : ViewModel() {
     private val _surveyWarning = MutableStateFlow(false)
     val surveyWarning: StateFlow<Boolean> = _surveyWarning.asStateFlow()
@@ -69,47 +62,9 @@ class DashboardViewModel @Inject constructor(
 
     suspend fun updateResourceNotification(userId: String?) {
         try {
-            databaseService.executeTransactionAsync { realm ->
-                val resourceCount = BaseResourceFragment.getLibraryList(realm, userId).size
-                if (resourceCount > 0) {
-                    val existingNotification = realm.where(RealmNotification::class.java)
-                        .equalTo("userId", userId)
-                        .equalTo("type", "resource")
-                        .findFirst()
-
-                    if (existingNotification != null) {
-                        existingNotification.message = "$resourceCount"
-                        existingNotification.relatedId = "$resourceCount"
-                    } else {
-                        createNotificationIfNotExists(realm, "resource", "$resourceCount", "$resourceCount", userId)
-                    }
-                } else {
-                    realm.where(RealmNotification::class.java)
-                        .equalTo("userId", userId)
-                        .equalTo("type", "resource")
-                        .findFirst()?.deleteFromRealm()
-                }
-            }
+            notificationRepository.updateResourceNotification(userId)
         } catch (e: Exception) {
             e.printStackTrace()
-        }
-    }
-
-    fun createNotificationIfNotExists(realm: Realm, type: String, message: String, relatedId: String?, userId: String?) {
-        val existingNotification = realm.where(RealmNotification::class.java)
-            .equalTo("userId", userId)
-            .equalTo("type", type)
-            .equalTo("relatedId", relatedId)
-            .findFirst()
-
-        if (existingNotification == null) {
-            realm.createObject(RealmNotification::class.java, "${UUID.randomUUID()}").apply {
-                this.userId = userId ?: ""
-                this.type = type
-                this.message = message
-                this.relatedId = relatedId
-                this.createdAt = Date()
-            }
         }
     }
